cloud_functions/send_daily_report/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.30.0
google-cloud-secret-manager==2.20.0
google-cloud-logging==3.11.0
sendgrid==6.11.0
jinja2==3.1.2
